refactor: 예약 카드 내부의 타입, 스타일을 컴포넌트에서 분리
