+++ frontend/src/features/book/Book.types.ts
@@ -1,0 +1,1 @@
+export type BookType = "pending" | "success" | "fail";
